Déploiement des 4 workers en production. force deploy

+++ libs/domains/accounting/config/delete-account-class/handler.ts
@@ -1,8 +1,14 @@
-import { type Db } from '@nba/db';
+import { type Db, AppError } from '@nba/db';
 import { DeleteAccountClassRepository } from './repository';
 import { DeleteAccountClassInput, DeleteAccountClassOutput } from "./dto";
 
 export async function deleteAccountClass(db: Db, code: DeleteAccountClassInput): Promise<DeleteAccountClassOutput> {
   const repo = new DeleteAccountClassRepository();
+  const isUsed = await repo.isAccountClassUsed(db, code);
+  
+  if (isUsed) {
+    throw new AppError('Impossible de supprimer cette classe comptable car elle est rattachée à une ou plusieurs catégories.', 403);
+  }
+  
   return repo.deleteAccountClass(db, code);
 }
